+++ src/emit/index.ts
@@ -22224,6 +22224,9 @@ class Emitter {
             if (key !== "encoding") {
                 unsupported(prop.name, `os.userInfo unsupported option ${key ?? ts.SyntaxKind[prop.name.kind]}`);
             }
+            if (this.isUndefinedExpression(prop.initializer)) {
+                continue;
+            }
             if (!ts.isStringLiteralLike(prop.initializer) || (prop.initializer.text !== "utf8" && prop.initializer.text !== "utf-8")) {
                 unsupported(prop.initializer, "os.userInfo currently supports UTF-8 encoding options only");
             }